Remove tcflush during BT_ON
During BT_On procedure when we call tcflush function, on kernel side
it did not update the UART Tx status, so that in the UART Tx callback
function, it just updated the TTY buffer write pointer and discarded
the change baud rate command. Fix it by remove the tcflush function
call in BT_On procedure.

CRs-fixed:1077744

Change-Id: Ibde0a038052ea5dff2fd4bdb6059d8e964d3cebd