BT: Ensure WCNSS_FILTER is killed as part of turn OFF
WCNSS_FILTER might go into bad state and not exit gracefully
in the allotted time during turn OFF. This lead to the
"wc_transport.start_hci" to stay as true and hence filter couldn’t
be restarted. Ensure it is killed so that next turn ON cycle is
not affected.

CRs-Fixed: 2055166
Change-Id: Iadd94079104a883d2e9e974631d664b0451be196